- Adjusted cmake build environment - Added test code to read in gtfs files - Added test code to transform gtfs files into a packed graph - Several small changes in Core - Added capability of accounting for turn penalties into the A star algorithm

+++ polaris_git/Libs/Routing/Graph_Pool.h
@@ -20,7 +20,7 @@ namespace polaris
 			return this_component()->Get_Graph(graph);
 		}
 
-		void Update_Edge(int id, boost::container::vector<int>* graph_set = nullptr, typename edge_update_callback<edge_type>::type callback = nullptr)
+		void Update_Edge(long long id, boost::container::vector<int>* graph_set = nullptr, typename edge_update_callback<edge_type>::type callback = nullptr)
 		{
 			this_component()->Update_Edge(id,graph_set,callback);
 		}
